+ Added `extraMenuItems` to `ViewManager` + Moved `parseMenuItems` to reusable util
