+++ desktop/cmp/button/AppMenuButton.ts
@@ -4,17 +4,13 @@
  *
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
-import {MenuItemProps} from '@blueprintjs/core';
-import {hoistCmp, isMenuItem, MenuItemLike, XH} from '@xh/hoist/core';
+import {hoistCmp, MenuItemLike, XH} from '@xh/hoist/core';
 import {ButtonProps, button} from '@xh/hoist/desktop/cmp/button';
 import '@xh/hoist/desktop/register';
 import {Icon} from '@xh/hoist/icon';
-import {menu, menuDivider, menuItem, popover} from '@xh/hoist/kit/blueprint';
-import {wait} from '@xh/hoist/promise';
-import {filterConsecutiveMenuSeparators, isOmitted} from '@xh/hoist/utils/impl';
+import {menu, popover} from '@xh/hoist/kit/blueprint';
+import {parseMenuItems} from '@xh/hoist/utils/impl';
 import {withDefault} from '@xh/hoist/utils/js';
-import {clone, isEmpty} from 'lodash';
-import {ReactNode} from 'react';
 
 export interface AppMenuButtonProps extends ButtonProps {
     /**
@@ -175,43 +171,4 @@ function buildMenuItems(props: AppMenuButtonProps) {
     ];
 
     return parseMenuItems([...extraItems, '-', ...defaultItems]);
-}
-
-function parseMenuItems(items: MenuItemLike[]): ReactNode[] {
-    items = items.map(item => {
-        if (!isMenuItem(item)) return item;
-
-        item = clone(item);
-        item.items = clone(item.items);
-        item.prepareFn?.(item);
-        return item;
-    });
-
-    return items
-        .filter(it => !isMenuItem(it) || (!it.hidden && !isOmitted(it)))
-        .filter(filterConsecutiveMenuSeparators())
-        .map(item => {
-            if (item === '-') return menuDivider();
-            if (!isMenuItem(item)) return item;
-
-            const {actionFn} = item;
-
-            // Create menuItem from config
-            const cfg: MenuItemProps = {
-                text: item.text,
-                icon: item.icon,
-                intent: item.intent,
-                className: item.className,
-                onClick: actionFn ? e => wait().then(() => actionFn(e)) : null, // do async to allow menu to close
-                disabled: item.disabled
-            };
-
-            // Recursively parse any submenus
-            if (!isEmpty(item.items)) {
-                cfg.children = parseMenuItems(item.items);
-                cfg.popoverProps = {openOnTargetFocus: false};
-            }
-
-            return menuItem(cfg);
-        });
 }
